split UI files
